Refactor into a function
